+++ src/main/kotlin/com/github/hirofumi/termback/TermbackSettings.kt
@@ -6,11 +6,23 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.SimplePersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
+import com.intellij.openapi.util.SystemInfo
+
+enum class NotificationDestination {
+    IDE,
+    SYSTEM,
+    ;
+
+    companion object {
+        val default: NotificationDestination = if (SystemInfo.isMac) SYSTEM else IDE
+    }
+}
 
 @Service(Service.Level.APP)
 @State(name = "com.github.hirofumi.termback.TermbackSettings", storages = [Storage("termback.xml")])
 class TermbackSettings : SimplePersistentStateComponent<TermbackSettings.State>(State()) {
     class State : BaseState() {
+        var notificationDestination by enum(NotificationDestination.default)
         var skipPopupWhenSingleNotification by property(false)
     }