+++ src/main/kotlin/com/github/hirofumi/termback/notification/channels/macos/NSUserNotificationCenter.kt
@@ -1,0 +1,6 @@
+package com.github.hirofumi.termback.notification.channels.macos
+
+import com.intellij.ui.mac.foundation.Foundation
+import com.intellij.ui.mac.foundation.ID
+
+internal fun notificationCenter(): ID = Foundation.invoke("NSUserNotificationCenter", "defaultUserNotificationCenter")